use ssyevd rather than cgeev for better numerical stabiity.
H has to be calculated as complex, but as we know it is an observable we can
cast it to real and treat it as a symmetric matrix. This allows for more
stable solvers to be used, and so we can get very good orthonormal vectors -
cgeev gave poorly orthonormalised vectors.